requirements.txt
absl-py==0.11.0
aiocontextvars==0.2.2
asn1crypto==0.22.0
# awscli==1.16.6
# awscli
backcall==0.1.0
bleach==3.0.2
boto==2.49.0
botocore==1.11.6
cached-property==1.5.2
cachetools==4.1.1
certifi==2017.7.27.1
cffi==1.14.0
chardet==3.0.4
cloudpickle==0.6.1
colorama==0.3.9
conda-pack==0.5.0
contextvars==2.4
cryptography==2.0.3
cycler==0.10.0
Cython==0.29.6
dask==0.20.2
dataclasses==0.8
decorator==4.3.0
docutils==0.14
easydict==1.8
entrypoints==0.2.3
future==0.18.2
# fvcore==0.1.3.post20210311
# fvcore
glog==0.3.1
google-auth==1.23.0
google-auth-oauthlib==0.4.2
grpcio==1.33.2
h5py==3.1.0
idna==2.6
imageio==2.4.1
immutables==0.14
importlib-metadata==3.1.0
iopath==0.1.4
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jmespath==0.9.4
joblib==0.17.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
loguru==0.5.3
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.0.0
mistune==0.8.3
mkl-fft==1.0.2
mkl-random==1.0.1.1
# mkl-service==2.3.0
msgpack==1.0.0
msgpack-numpy==0.4.7.1
multimethod==1.4
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.2
notebook==5.7.2
# numpy==1.19.4
oauthlib==3.1.0
opencv-python==3.4.2.17
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==8.0.1
Pillow-SIMD==5.3.0.post0
portalocker==2.2.1
prometheus-client==0.4.2
prompt-toolkit==2.0.7
protobuf==3.14.0
psutil==5.7.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.2.0
pyOpenSSL==17.2.0
pyparsing==2.2.2
PySocks==1.6.7
python-dateutil==2.8.0
python-gflags==3.1.2
PyWavelets==1.0.1
# pyyaml
PyYAML==5.4.1
pyzmq==17.1.2
qtconsole==4.4.2
requests==2.25.0
requests-oauthlib==1.3.0
rsa==3.4.2
ruamel.yaml==0.11.14
s3transfer==0.1.13
scikit-image==0.14.1
scikit-learn==0.23.2
scipy==1.5.4
Send2Trash==1.5.0
sklearn==0.0
tabulate==0.8.7
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorpack==0.10.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
toolz==0.9.0
# torch==1.1.0
torchpack==0.2.2
# torchvision==0.2.1
tornado==5.1
tqdm==4.53.0
traitlets==4.3.2
typing-extensions==3.7.4.3
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.4.2
yacs==0.1.8
zipp==3.4.0
